requirements.txt
apache-airflow[google]==2.5.3